% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{adjust_ancestry}
\alias{adjust_ancestry}
\title{Fix ancestry information.}
\usage{
adjust_ancestry(graph)
}
\arguments{
\item{graph}{a graph as returned by \link{traverse}.}
}
\value{
Input \code{graph} with \code{children} and \code{parents} adjusted to
contain only identifiers of nodes present in \code{graph}.
}
\description{
After defining a graph subset by a traverse, it might be desirable
to have the \code{children} and \code{parents} keys of each graph node contain
only identifiers of nodes present in the graph.
}
\seealso{
ancestry_graph, traverse
}
