% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-specialized.R
\name{ancestry_graph}
\alias{ancestry_graph}
\title{Build a connected ancestry graph.}
\usage{
ancestry_graph(chosen_ids, all_ids, store)
}
\arguments{
\item{store}{object store; see \link[storage:object_store]{storage::object_store}}

\item{ids}{identifiers of objects in \code{store}}
}
\value{
A \code{list} named according to \code{ids}; each element is a list with
two keys: \code{parents} and \code{children}.
}
\description{
For each identifier in \code{ids}, \code{ancestry_graph} creates and returns a
\code{list} with the two following keys:
}
\details{
\itemize{
\item \code{parents} which is read directly from \code{store}
\itemize{
\item \code{children} which is inferred from \code{parents}
}
}

Those lists are wrapped in a single list and named with values from \code{ids}.
Furthermore, \code{ancestry_graph} verifies that all values present in \code{parents}
and \code{children} are also present in \code{ids} and that the resulting graph is
connected, that is, whether there is a path between any pair of nodes.
}
