% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact.R
\name{replot}
\alias{replot}
\title{Re-plot an archived plot.}
\usage{
replot(x, method = "re-evaluate")
}
\arguments{
\item{x}{plot artifact, as returned by \code{\link[=read_artifacts]{read_artifacts()}}.}

\item{method}{\code{"replay"} or \code{"re-evaluate"}}
}
\description{
There are two ways of re-creating the plot. One (\code{method == "replay"})
is to call \link[grDevices:replayPlot]{grDevices::replayPlot}. The other (\code{method == "re-evaluate"})
is to restore the state of R session at the time of plotting and re-run
the expression that created the original plot.
}
