% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact.R
\name{is_artifact}
\alias{is_artifact}
\alias{artifact_store}
\alias{artifact_is}
\alias{artifact_data}
\alias{artifact_commit}
\title{Manipulating and processing artifacts.}
\usage{
is_artifact(x)

artifact_store(x)

artifact_is(x, what)

artifact_data(x)

artifact_commit(x)
}
\arguments{
\item{x}{object to be tested; \code{artifact} to be processed.}

\item{what}{property of an \code{artifact} to be tested.}
}
\description{
\code{is_artifact} tests whether an object is an \code{artifact}.

\code{artifact_store} gives access to \link[storage:object_store]{storage::object_store}
where artifact \code{x} is stored.

\code{artifact_is} answers various questions about an
artifact.

\code{artifact_data} loads the actual artifact object. The
output might be large and thus it is loaded lazily only upon explicit
request.

\code{artifact_commit} returns the parent \code{commit} for the
given artifact.
}
\examples{
\dontrun{
artifact_is(a, 'plot')
}
}
\seealso{
replot
}
