% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{binding_sites}
\alias{binding_sites}
\title{binding_sites}
\usage{
binding_sites(
  p_value_matrix,
  effect_size_matrix,
  p_threshold = 0.1,
  width_threshold = 10
)
}
\arguments{
\item{p_value_matrix}{matrix of adjusted pvalues, where each row is a base pair position and each column is a center window.}

\item{effect_size_matrix}{matrix of effect sizes, where each row is a base pair position and each column is a center window.}

\item{width_threshold}{minimum width reported binding sites}

\item{pval_threshold}{fdr of reported binding sites}
}
\value{
a list of non overlapping footprints with maximum pval, but the positions and sizes of reported footprints are decided by effect size
}
\description{
searching binding sites by pval
}
\details{
the positions and sizes of reported footprints are decided by effect size
}
