% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{binding_sites_pval}
\alias{binding_sites_pval}
\title{binding_sites_pval}
\usage{
binding_sites_pval(
  p_value_matrix,
  smoothed = FALSE,
  pval_threshold = 0.1,
  width_threshold = 10
)
}
\arguments{
\item{p_value_matrix}{matrix of adjusted pvalue, where each row is a base pair position and each column is a center window.}

\item{smoothed}{T/F, if the pvalue needed to be smoothed within neighbourhood}

\item{pval_threshold}{fdr of reported binding sites}

\item{width_threshold}{minimum width reported binding sites}
}
\value{
a list of non overlapping footprints with maximum log pvalues, rank by log pvalues
}
\description{
searching binding sites by pvalue
}
\details{
searching binding sites by pvalue
}
