% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{new_commit}
\alias{new_commit}
\alias{commit_store}
\alias{is_valid_commit}
\alias{as_environment}
\title{Creating and manipulating commits.}
\usage{
new_commit(id, store)

commit_store(x)

is_valid_commit(x)

as_environment(x, parent = new.env())
}
\arguments{
\item{id}{commit identifier.}

\item{store}{object store, see \link[storage:object_store]{storage::object_store}.}

\item{x}{commit object.}

\item{parent}{parent environment for the newly created one.}
}
\description{
Creating and manipulating commits.
}
