% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{commit}
\alias{commit}
\alias{is_commit}
\alias{print.commit}
\alias{commit_data}
\alias{commit_checkout}
\title{Commit API.}
\usage{
is_commit(x)

\method{print}{commit}(x, ...)

commit_data(x)

commit_checkout(commit, env)
}
\arguments{
\item{x}{commit object.}

\item{...}{Arguments passed on to \code{base::print}
\describe{
  \item{x}{an object used to select a method.}
}}

\item{commit}{as returned by \link{read_commits}.}

\item{env}{Environment to check-out artifacts to.}
}
\description{
Commit API.

Copy commit into environment.
}
