% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{commit_checkout}
\alias{commit_checkout}
\title{Copy commit into environment.}
\usage{
commit_checkout(commit, env)
}
\arguments{
\item{commit}{as returned by \link{read_commits}.}

\item{env}{Environment to check-out artifacts to.}
}
\description{
Copy commit into environment.
}
