% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{commit_write}
\alias{commit_write}
\title{Write commit to an object store.}
\usage{
commit_write(.)
}
\arguments{
\item{store}{An object store, e.g. \link[storage:filesystem]{storage::filesystem}.}

\item{commit}{A \link{commit} object.}
}
\description{
An assumption is made that all objects are already stored and only
the commit itself needs to be written to the object store.
}
