% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeCountTensor}
\alias{computeCountTensor}
\title{computeCountTensor}
\usage{
computeCountTensor(
  frags,
  regions,
  barcodeGroups,
  chunkSize = 2000,
  nCores = 2,
  pairedEnd = TRUE
)
}
\arguments{
\item{regions}{Genomic ranges of the regions to footprint}

\item{barcodeGroups}{Data.frame specifying membership of barcodes in pseudobulks. First column is barcodes and second is groupID}

\item{chunkSize}{Chunk size for parallel processing of regions}

\item{nCores}{Number of cores to use}

\item{pairedEnd}{Whether data is paired-end or single-end}

\item{frags_path}{path to frags}
}
\value{
A count tensor file
}
\description{
help function for get insertion count for each base pair
}
\details{
help function for get insertion count for each base pair from frag file
}
