% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{finetuned_model}
\alias{finetuned_model}
\title{finetuned_model}
\usage{
finetuned_model(
  obsbias_path,
  PRINT_model_path,
  finetuned_model_save_path,
  finetuned_model_name,
  nCores = 2
)
}
\arguments{
\item{obsbias_path}{path to file of observed bias in mito DNA regions}

\item{PRINT_model_path}{path to PRINT model}

\item{finetuned_model_save_path}{path to save the finetuned model}

\item{finetuned_model_name}{name to save the finetuned model}

\item{nCores}{Number of cores to use}
}
\value{
A finetuned model
}
\description{
Finetune PRINT model by observed bias
}
\details{
Finetune PRINT model by observed bias in mito DNA regions
}
