% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fisher_comparison}
\alias{fisher_comparison}
\title{fisher_comparison}
\usage{
fisher_comparison(
  binding_site,
  Tn5Insertion_1,
  Tn5Insertion_2,
  pred_bias,
  positions,
  flank_window_range = 40,
  flank_step = 2,
  alternative = "two.sided"
)
}
\value{
a list of footprint binding sites with pval indicating significance of differential bindings between two conditions
}
\description{
compare binding events between different conditions with fisher exact test
}
\details{
keep for legacy
}
