% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository-impl.R
\name{format_expr}
\alias{format_expr}
\title{Pretty-format an R expression.}
\usage{
format_expr(expr, indent = "  ")
}
\arguments{
\item{expr}{Expression.}

\item{indent}{Indentation for all lines of the expression.}
}
\description{
Internally calls \code{\link[styler:style_text]{styler::style_text()}} but adds line breaks after each
occurence of the pipe operator.
}
