% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bardata}
\alias{get_bardata}
\title{get_bardata}
\usage{
get_bardata(counts, regionsBed, bias, i, groupIDs)
}
\arguments{
\item{counts}{the count tensor from get_count}

\item{regionsBed}{a data frame for peak, containing columns chr, start. end}

\item{bias}{the bias tensor from get_bias}

\item{i}{ID of peak of interest in regionsBed file}

\item{groupIDs}{group(s) of interest}
}
\value{
A data frame containing position, insertions and predicted bias
}
\description{
Create the bias and insertions data for the peak
}
\details{
create the bias and insertions data for the peak
}
