% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bias}
\alias{get_bias}
\title{get_bias}
\usage{
get_bias(
  regionsBed,
  referenceGenome = "hg38",
  genome = NULL,
  nCores = 2,
  chunkSize = 2000,
  model_use = "Tn5_NN_model.h5",
  model_path = "../../",
  save_path = NULL
)
}
\arguments{
\item{regionsBed}{a data frame for peak, containing columns chr, start. end}

\item{referenceGenome}{hg19/hg38/mm10/rn7}

\item{genome}{Full genome sequences if not hg19/hg38/mm10/rn7}

\item{nCores}{Number of cores to use}

\item{chunkSize}{Chunk size for parallel processing of regions}

\item{model_use}{which bias model to use}

\item{model_path}{the path to model}

\item{save_path}{the path to save the bias file}
}
\value{
A bias tensor
}
\description{
Get bias for each base pair
}
\details{
Get bias for each base pair by finetuned/trained model
}
