% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_count}
\alias{get_count}
\title{get_count}
\usage{
get_count(
  frags_path,
  regionsBed,
  barcodeGroups = NULL,
  chunkSize = 2000,
  nCores = 2,
  pairedEnd = TRUE,
  startsAre0based = TRUE
)
}
\arguments{
\item{frags_path}{path to frags}

\item{regionsBed}{a data.frame for peak, containing columns chr, start. end}

\item{barcodeGroups}{a data.frame for barcodes in pseudobulks. First column is barcodes and second is group info}

\item{chunkSize}{Chunk size for parallel processing of regions}

\item{nCores}{Number of cores to use}

\item{pairedEnd}{Whether data is paired-end or single-end}

\item{startsAre0based}{Whether the coordinates are 0-based}
}
\value{
A count tensor file
}
\description{
get insertion count for each base pair
}
\details{
get insertion count for each base pair from frag file
}
