% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_insertions}
\alias{get_insertions}
\title{get_insertions}
\usage{
get_insertions(countData, regionInd, groupIDs, width)
}
\arguments{
\item{countData}{Tn5 insertion count on a particular region from get_count}

\item{groupIDs}{Group name}

\item{width}{Width of the region}
}
\value{
A vector containing insertions for every base pair
}
\description{
For given group, get insertions
}
\details{
For given group, get insertions from count data on a particular region
}
