% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChunkInterval}
\alias{getChunkInterval}
\title{getChunkInterval}
\usage{
getChunkInterval(x, chunkSize = 2000)
}
\arguments{
\item{x}{Vector or list}

\item{chunkSize}{Size of a single chunk}
}
\value{
start and end of x after chunking with chunkSize
}
\description{
Chunk a vector/list x into chunks. Return starts and ends of chunks
}
\details{
Chunk a vector/list x into chunks. Return starts and ends of chunks
}
