% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_of_artifacts}
\alias{graph_of_artifacts}
\title{Turn a list of artifacts into a graph of artifacts.}
\usage{
graph_of_artifacts(artifacts, store)
}
\description{
A list of artifacts, whose \code{names()} are artifact identifiers, is
transformed into a graph by adding a \code{children} key each element of
the list, removing \code{parents} that do exist in the list and finally
re-assigning \code{parents} if a certain parent artifact is not in the
input list.
}
\details{
two extra keys to each element
of the input list: \code{parents} and \code{children}
}
