% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{history_ancestors}
\alias{history_ancestors}
\alias{history_ends}
\alias{history_match}
\title{Show ancestors.}
\usage{
history_ancestors(x, id)

history_ends(x)

history_match(x, m)
}
\description{
Show ancestors.
}
