% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadesc.R
\name{description}
\alias{description}
\title{Provide a summary of an object.}
\usage{
description(tags)
}
\arguments{
\item{tags}{Tags retrieved from an object store.}
}
\description{
Provide a summary of an object.
}
