% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{is_query}
\alias{is_query}
\title{Identify \code{query} objects.}
\usage{
is_query(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"query"}.
}
\description{
Identify \code{query} objects.
}
