% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{london_meters}
\alias{london_meters}
\title{Work with a sample artifact repository.}
\usage{
london_meters()
}
\description{
Returns a handle to a read/write copy of an artifact repository
installed with the package.
}
