% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mito_downsamples}
\alias{mito_downsamples}
\title{mito_downsamples}
\usage{
mito_downsamples(
  mito_counts,
  mito_bias,
  mito_regions,
  i,
  average_insertions,
  nCore = 2
)
}
\arguments{
\item{mito_counts}{the count tensor from get_count on mito regions}

\item{mito_bias}{the bias tensor from get_bias on mito regions}

\item{mito_regions}{a data frame for mito regions, containing columns chr, start. end}

\item{i}{ID of peak of interest in mito region, has to be the test region}

\item{average_insertions}{average insertions of interest}

\item{nCore}{fdr}
}
\value{
A data frame containing position, width, pvalue for non overlapping 'footprintings' (not necessarily significant)
}
\description{
Get footprints on downsampled mito DNA regions with all given average insertions
}
\details{
Get footprints on downsampled mito DNA regions with all given average insertions
}
