% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mito_fdr}
\alias{mito_fdr}
\title{mito_fdr}
\usage{
mito_fdr(
  mito_counts,
  mito_bias,
  mito_regions,
  i,
  seeds,
  average_insertions,
  alpha = 0.05,
  nCore = 2
)
}
\arguments{
\item{mito_counts}{the count tensor from get_count on mito regions}

\item{mito_bias}{the bias tensor from get_bias on mito regions}

\item{mito_regions}{a data frame for mito regions, containing columns chr, start. end}

\item{i}{ID of peak of interest in mito region, has to be the test region}

\item{seeds}{random seed for set.seed()}

\item{average_insertions}{average insertions of interest}

\item{alpha}{fdr}

\item{nCore}{Number of cores to use}
}
\value{
A data frame containing coverage, fdr threshold
}
\description{
Create thresholds for False Discovery Control, stratified by average insertions
}
\details{
Create thresholds for stratified False Discovery Control based on mito regions
}
