% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{NB_footprintings}
\alias{NB_footprintings}
\title{NB_footprintings}
\usage{
NB_footprintings(
  Tn5Insertion,
  pred_bias,
  positions = 1:length(Tn5Insertion),
  center_sizes = 2:100,
  flank_window_range = 40,
  flank_step = 4,
  p.adjust.method = "BH",
  wide_flank = TRUE,
  nCores = 2
)
}
\arguments{
\item{Tn5Insertion}{a vector of Tn5 insertion for consecutive positions in the peak, matched with pred_bias, positions}

\item{pred_bias}{a vector of predicted Tn5 bias for consecutive positions in the peak, matched with Tn5Insertion, positions}

\item{positions}{a vector of consecutive positions in chr, matched with Tn5Insertion, pred_bias}

\item{center_sizes}{a vector of radium sizes of footprint}

\item{flank_step}{the step size of flanking sizes seq}

\item{p.adjust.method}{BH or BY, pval adjusted method}

\item{wide_flank}{if the flanking is 2 times of not}

\item{nCore}{Number of cores to use}
}
\value{
a list of adjusted pval and effect size matrix, where each row is a base pair position and each column is a center window.
}
\description{
calculate footprintings by negative binomial method,
}
\details{
calculate pvalue and effect sizes of footprintings by negative binomial method,
}
