% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_insertions}
\alias{plot_insertions}
\title{plot_multiscale_logp}
\usage{
plot_insertions(Tn5Insertion, positions = 1:length(Tn5Insertion), chr = NULL)
}
\arguments{
\item{Tn5Insertion}{a vector of Tn5 insertions per base pair}

\item{positions}{a vector of positions of the peak of interest}

\item{chr}{chromosome of the peak of interest}
}
\value{
a ggplot object of insertions per base pair
}
\description{
visualize insertions
}
\details{
visualize insertions per base pair
}
