% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_multiscale_effectsize}
\alias{plot_multiscale_effectsize}
\title{plot_multiscale_effectsize}
\usage{
plot_multiscale_effectsize(
  effect_size_matrix,
  p_value_matrix,
  chr = NULL,
  scale = FALSE
)
}
\arguments{
\item{effect_size_matrix}{matrix of effect size, where each row is a base pair position and each column is a center window.}

\item{p_value_matrix}{matrix of adjusted p value, where each row is a base pair position and each column is a center window.}

\item{chr}{chromosome of the peak of interest}

\item{scale}{T/F if the scale of effect size min at 0 and pvalue max at 0.9}

\item{smooth}{T/F if the -log10 pval needs to be smoothed}
}
\value{
a ggplot object of multiscale effect size per base pair
}
\description{
visualize multiscale effect size
}
\details{
visualize multiscale effect size per base pair
}
