% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_multiscale_logp}
\alias{plot_multiscale_logp}
\title{plot_multiscale_logp}
\usage{
plot_multiscale_logp(p_value_matrix, chr = NULL, scale = FALSE, smooth = T)
}
\arguments{
\item{p_value_matrix}{matrix of adjusted pvalues, where each row is a base pair position and each column is a center window.}

\item{chr}{chromosome of the peak of interest}

\item{scale}{T/F if the scale of -log10 pval max at 2 (pval = 0.01)}

\item{smooth}{T/F if the -log10 pval needs to be smoothed}
}
\value{
a ggplot object of multiscale -log10 pval per base pair
}
\description{
visualize multiscale -log10 pval
}
\details{
visualize multiscale -log10 pval per base pair
}
