% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository-impl.R
\name{plot_as_svg}
\alias{plot_as_svg}
\alias{plot_as_png}
\alias{svg_equal}
\alias{png_equal}
\title{Returns a base64-encoded, SVG plot.}
\usage{
plot_as_svg(pl)

plot_as_png(pl)

svg_equal(a, b)

png_equal(a, b)
}
\arguments{
\item{pl}{Plot recorded by \code{\link[=recordPlot]{recordPlot()}}.}

\item{a}{First SVG image.}

\item{b}{Second SVG image.}

\item{pl}{Plot recorded by \code{\link[=recordPlot]{recordPlot()}}.}
}
\value{
\code{character} string, base64-encoded SVG plot.

\code{character} string, base64-encoded PNG plot.

\code{TRUE} if SVGs are the same plot-wise.
}
\description{
Returns a base64-encoded, SVG plot.

Returns a base64-encoded, PNG plot.

SVG images are processed in this package as base64-encoded, XML text
data. When produced, certain differences are introduced in XML
attributes that have no effect on the final plots. That is why,
however, SVG plots need to be compared graphically and not textually.
This function produces a thumbnail of each SVG image and then
compares the raster graphic.
}
