% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_cluster}
\alias{prep_cluster}
\title{prep_cluster}
\usage{
prep_cluster(len, n_cores = 2)
}
\arguments{
\item{len}{Number of elemnts in the iterable list}

\item{n_cores}{Number of cores to use}
}
\value{
opts and cl for foreach
}
\description{
Prepares cluster for parallel computing using foreach
}
\details{
Prepares cluster for parallel computing using foreach
}
