% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{tag_names}
\alias{tag_names}
\alias{tag_values}
\title{Extra query API.}
\usage{
tag_names(x)

tag_values(x)
}
\description{
\code{tag_names} returns all tag names occuring among objects
selected in the query \code{x}.

\code{tag_values} returns a \code{list} of vectors of values for
all tags occuring among objects selected in the query \code{x}.
}
