% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{new_query}
\alias{new_query}
\alias{as_query}
\title{Internal API for query objects.}
\usage{
new_query(x)

as_query(x)
}
\arguments{
\item{x}{\code{repository} object or an object to be turned into a \code{query}.}
}
\description{
Internal API for query objects.
}
