% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{is_query}
\alias{is_query}
\alias{print.query}
\alias{tag_names}
\alias{tag_values}
\alias{unselect}
\alias{execute}
\title{Identify \code{query} objects.}
\usage{
is_query(x)

\method{print}{query}(x, ..., simplify = FALSE)

tag_names(x)

tag_values(x)

unselect(.data)

execute(x, .warn = TRUE)
}
\arguments{
\item{x}{Object to be tested.}

\item{...}{further arguments passed to or from other methods.}

\item{simplify}{simplify output formatting.}

\item{.data}{\code{query} object.}

\item{.warn}{print warnings.}

\item{x}{\code{query} object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"query"}.
}
\description{
Identify \code{query} objects.

\code{tag_names} returns all tag names occuring among objects
selected in the query \code{x}.

\code{tag_values} returns a \code{list} of vectors of values for
all tags occuring among objects selected in the query \code{x}.

\code{unselect} clears the list of selected tag names.

\code{execute} runs the query and retrieves its results.
}
