% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_session_id}
\alias{r_session_id}
\title{Generate and return an unique session id.}
\usage{
r_session_id()
}
\description{
Session id is stored via \code{options()} in order to preserve it regardless
of the changes in values coming as input, namely changes in the output
of \code{sessionInfo()}.
}
