% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rawplot}
\alias{rawplot}
\alias{print.rawplot}
\alias{plot.rawplot}
\title{Container for plots.}
\usage{
rawplot(rp)

\method{print}{rawplot}(x, ...)

\method{plot}{rawplot}(x, ...)
}
\description{
Container for plots.
}
