% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{arrange}
\alias{summarise}
\alias{n}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{n}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

