% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{remove_intersection}
\alias{remove_intersection}
\title{remove_intersection}
\usage{
remove_intersection(multiscale_data, binding_sites)
}
\arguments{
\item{multiscale_data}{a vector of Tn5 insertion for consecutive positions in the peak, matched with pred_bias, positions}

\item{binding_sites}{a vector of predicted Tn5 bias for consecutive positions in the peak, matched with Tn5Insertion, positions}
}
\value{
regions with pvalue non-overlapped with reported binding_sites
}
\description{
remove regions that overlap with the reported binding_sites
}
\details{
remove regions that overlap with the reported binding_sites
}
