% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{replot_as}
\alias{replot_as}
\alias{as_png}
\alias{as_svg}
\title{Returns a base64-encoded, SVG or PNG plot.}
\usage{
replot_as(pl, type, ...)

as_png(x, width = 1280, height = 720)

as_svg(x)
}
\arguments{
\item{pl}{Plot recorded by \code{\link[grDevices:recordPlot]{grDevices::recordPlot()}}.}

\item{type}{Output format: \code{"png"} or \code{"svg"}.}

\item{...}{Passed to \code{\link[=svg]{svg()}} or \code{\link[=png]{png()}}.}

\item{x}{\code{rawplot} object.}

\item{width}{PNG width.}

\item{height}{PNG height.}
}
\value{
\code{character} string, base64-encoded SVG or PNG plot.
}
\description{
Returns a base64-encoded, SVG or PNG plot.
}
