% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\docType{package}
\name{repo-package}
\alias{repo}
\alias{repo-package}
\title{Repo: The Data-centered Data Flow Manager}
\description{
A data manager meant to avoid manual storage/retrieval of data to/from the file system. It builds one (or more) centralized repository where R objects are stored with rich annotations, including corresponding code chunks, and easily searched and retrieved. See Napolitano (2017) \doi{10.1186/s12859-017-1510-6} for further information.
}
\details{
The Repo package is meant to help with the management of R data
files. It builds one (or more) centralized repository where R
objects are stored together with corresponding annotations, tags,
dependency notes, provenance traces. It also provides navigation
tools to easily locate and load previously stored resources.

Create a new repository with \code{rp <- repo_open()}.

Given the object rp of class \code{repo}, the \code{repo} command
\code{foo} must be called like this: \code{rp$foo()}. However, the
public name of \code{foo} will be \code{repo_foo}, and this name
must be used to get help (\code{?repo_foo}).

For a complete list of functions, use \code{library(help = "repo")}.
}
\author{
Francesco Napolitano \email{franapoli@gmail.com}
}
\keyword{internal}
