% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_attr}
\alias{repo_attr}
\title{Get item attribute.}
\usage{
repo_attr(rp, name, attrib)
}
\arguments{
\item{rp}{An object of class repo.}

\item{name}{An item name.}

\item{attrib}{An attribute name (currently can be only "path").}
}
\value{
The item's attribute value.
}
\description{
Get item attribute.
}
\examples{
rp_path <- file.path(tempdir(), "example_repo")
rp <- repo_open(rp_path, TRUE)
rp$put(1, "item1", "Sample item 1", "tag1")
print(rp$attr("item1", "path"))

## wiping temporary repo
unlink(rp_path, TRUE)
}
\seealso{
repo_entries, repo_get
}

