% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_bulkedit}
\alias{repo_bulkedit}
\title{Edit all items info using a text file.}
\usage{
repo_bulkedit(repo, outfile = NULL, infile = NULL)
}
\arguments{
\item{repo}{An object of class repo.}

\item{outfile}{Name of a file to put entries data to.}

\item{infile}{Name of a file to read entries data from.}
}
\value{
Used for side effects.
}
\description{
Edit all items info using a text file.
}
\details{
Exactly one of `outfile` or `infile` must be supplied. All repo entries data are copied to a tab-separated file when using the `outfile` parameter. All repo entries are updated reading from `infile` when the `infile` parameter is used. Within the TAGS field, tags must be comma-separated. The system writes a checksum to the `outfile` that precludes using it as `infile` if repo has changed in the meantime.
}
\examples{
repo_path <- file.path(tempdir(), "example_repo")
repo <- repo_open(repo_path, TRUE)
repo$put(1, "item1", "Sample item 1", c("tag1", "tag2"), replace=TRUE)

items_data_file <- tempfile()
repo$bulkedit(items_data_file)
## Manually edit items_data_file, then update items:
repo$bulkedit(infile=items_data_file)

## wiping temporary repo
unlink(repo_path, TRUE)
}
\seealso{
repo_set
}

