% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_public.R
\name{repo_chunk}
\alias{repo_chunk}
\title{Shows code chunk associated with an item}
\usage{
repo_chunk(name)
}
\arguments{
\item{name}{Item name.}
}
\value{
List of lines of code, invisibly.
}
\description{
Shows code chunk associated with an item
}
