% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_cpanel.R
\name{repo_cpanel}
\alias{repo_cpanel}
\title{Open a visual interface to the repo}
\usage{
repo_cpanel(reporoot = NULL, env = globalenv())
}
\arguments{
\item{reporoot}{An object of class repo. Can be NULL like for repo_open.}

\item{env}{Environment to export variables to. Defaults to globalenv.}
}
\value{
Used for side effects.
}
\description{
Opens a browser window with a Shiny interface to a repo. The
interface is preliminary and has some exploration features together
with a "Load into workspace" button for a selected item.
}
