% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_public.R
\name{repo_depends}
\alias{repo_depends}
\title{Returns item's dependencies}
\usage{
repo_depends(name)
}
\arguments{
\item{name}{The name of a repository item.}
}
\value{
The items on which the input item depends.
}
\description{
Returns item's dependencies
}
