% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_entries}
\alias{repo_entries}
\title{Low-level list of item entries.}
\usage{
repo_entries(repo)
}
\arguments{
\item{repo}{An object of class repo.}
}
\value{
A list of item entries.
}
\description{
Low-level list of item entries.
}
\examples{
l <- repo_entries(repo)
}

