% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_find}
\alias{repo_find}
\title{Match items by matching any field}
\usage{
repo_find(x, find = NULL, all = F, show = "ds")
}
\arguments{
\item{x}{An object of class repo.}

\item{find}{Character to match any filed (see Details).}

\item{all}{Show also items tagged with "hide".}

\item{show}{Select columns to show.}
}
\value{
Used for side effects.
}
\description{
Match items by matching any field
}
\details{
This function actually calls print specifying the find
parameters. The find parameter can be any character string to be
matched against any item field, including string-converted size
(like "10x3").
}
\examples{
repo_path <- file.path(tempdir(), "example_repo")

repo <- repo_open(repo_path, TRUE)
repo$put(1, "item1", "Sample item 1", c("tag1", "tag2"), replace=TRUE)
repo$put(2, "item2", "Sample item 2", c("tag1", "hide"), replace=TRUE)
repo$put(3, "item3", "Sample item 3", c("tag2", "tag3"), replace=TRUE)
repo$print()
repo$find("tEm2")
repo$find("ag2", show="t")

## wiping the temp repo
unlink(repo_path, TRUE)

## wiping temporary repo
unlink(repo_path, TRUE)
}

