% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_get}
\alias{repo_get}
\title{Retrieve an item from the repo.}
\usage{
repo_get(repo, name)
}
\arguments{
\item{repo}{An object of class repo.}

\item{name}{An item's name.}
}
\value{
The previously stored object.
}
\description{
Retrieve an item from the repo.
}
\examples{
x <- repo_get(repo, "item1")
}

