% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_handlers}
\alias{repo_handlers}
\title{Provides alternative access to repo items.}
\usage{
repo_handlers(repo)
}
\arguments{
\item{repo}{An object of class repo.}
}
\value{
A list of functions.
}
\description{
Provides alternative access to repo items.
}
\examples{
repo <- repo_test()
h <- repo_handlers(repo)
## handlers have the same names as the items in the repo.
names(h)
## Without arguments, function "item1" loads item named "item1".
i1 <- h$item1()
## Arguments can be used to call other repo functions on the item.
h$item1("info")
}

