% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_public.R
\name{repo_has}
\alias{repo_has}
\title{Check wether a repository has an item}
\usage{
repo_has(name)
}
\arguments{
\item{name}{Item name.}
}
\value{
TRUE if \code{name} is in the repository, FALSE otherwise.
}
\description{
Check wether a repository has an item
}
