% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_public.R
\name{repo_options}
\alias{repo_options}
\title{Set repository-wide options}
\usage{
repo_options(...)
}
\arguments{
\item{...}{options to set}
}
\value{
if optional parameters are not passed, the current options
    are returned
}
\description{
Set repository-wide options
}
