% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_public.R
\name{repo_project}
\alias{repo_project}
\title{Defines and \code{put}-s a \code{project} item.}
\usage{
repo_project(name, description, replace = T)
}
\arguments{
\item{name}{character containing the name of the project}

\item{description}{character containing a longer description of the
project}

\item{replace}{logical, if T then an existing project item by the
same name will be overwritten.}
}
\value{
Used for side effects.
}
\description{
A \code{project} item is a special item containing session
information, including package dependencies. Every time a new item
is stored in the repository, it will automatically be assigned to
the current project, if one has been defined, and session
information will be updated.
}
