% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_pull}
\alias{repo_pull}
\title{Download item remote content}
\usage{
repo_pull(repo, name, replace = F)
}
\arguments{
\item{repo}{An object of class repo.}

\item{name}{Name of the existing item that will be updated.}

\item{replace}{If TRUE, existing item's object is overwritten.}
}
\value{
Used for side effects.
}
\description{
Download item remote content
}
\details{
Repo index files can be used as pointers to remote
    data. The pull function will download the actual data from the
    Internet, including regular items or attachment. Another use of
    the URL item's parameter is to attach a remote resource without
    downloading it.
}
\examples{
## Repository creation (or opening, if exists)
repo_path <- file.path(tempdir(), "example_repo")
repo <- repo_open(repo_path, TRUE)

## The following item will have remote source
repo$put("Local content", "item1", "Sample item 1", "tag",
         URL="http://www.francesconapolitano.it/repo/remote")
print(repo$get("item1"))
repo$pull("item1", replace=T)
print(repo$get("item1"))

## wiping temporary repo
unlink(repo_path, TRUE)
}

