% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_put}
\alias{repo_put}
\title{Create a new item in the repo.}
\usage{
repo_put(repo, obj, name, description, tags, src = NULL, replace = F)
}
\arguments{
\item{repo}{An object of class repo.}

\item{obj}{An R object to store in the repo.}

\item{name}{A character identifier for the new item.}

\item{description}{A character description of the item.}

\item{tags}{A list of tags to sort the item. Tags are useful for
selecting sets of items and run bulk actions.}

\item{src}{The item's provenance as a list of character. Usually
the name of the script producing the stored object, a website where
the object was downloaded, and so on. If one of the provenance
strings matches the name of a repo's item, this will create a
dependency link.}

\item{replace}{If the item exists, overwrite the specified fields.}
}
\value{
Used for side effects.
}
\description{
Create a new item in the repo.
}
\examples{
###code
}

