% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_public.R
\name{repo_related}
\alias{repo_related}
\title{Finds all items related to a set of item}
\usage{
repo_related(names, type = "all", excludeseed = F)
}
\arguments{
\item{names}{A \code{character} vector of item names.}

\item{type}{Can be one of "all", "to", "from". "to" recursively
finds items that \code{names} is attached to. "from" recursively
finds items that \code{names} depends on or is generated
by. "all" finds both (connected components including \code{names}.}

\item{excludeseed}{logical. If set to FALSE \code{names} will be
not included in the output list.}
}
\value{
A \code{character} vector of item names.
}
\description{
Relations are defined as in the dependency graph.
}
\seealso{
dependencies
}
