% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_stash}
\alias{repo_stash}
\title{Quickly store temporary data}
\usage{
repo_stash(repo, name, env = parent.frame())
}
\arguments{
\item{repo}{An object of class repo.}

\item{name}{A character containing the name of the variable to store.}

\item{env}{Environment containing the variable by the specified
name. Resolves to parent frame by default.}
}
\value{
Used for side effects.
}
\description{
Quickly store temporary data
}

