% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_stashclear}
\alias{repo_stashclear}
\title{Remove stashed data}
\usage{
repo_stashclear(repo)
}
\arguments{
\item{repo}{An object of class repo.}
}
\value{
Used for side effects.
}
\description{
Remove stashed data
}

