% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_sys}
\alias{repo_sys}
\title{Run system call on item}
\usage{
repo_sys(repo, name, command)
}
\arguments{
\item{repo}{An object of class repo.}

\item{name}{Name of a repo item. The path to the file that contains
the item will be passed to the system program.}

\item{command}{Shell command}
}
\value{
Used for side effects.
}
\description{
Run system call on item
}
\examples{
repo <- repo_open("_temp_repo", T)
## \dontrun {
## ## Creating a PDF file with a figure.
## pdf("afigure.pdf")
## ## Drawing a random plot in the figure
## plot(runif(100), runif(100))
## dev.off()
## ## Attaching the PDF file to the repo
## repo$attach("afigure.pdf", "A plot of random numbers", "repo_sys", replace=TRUE)
## ## don't need the PDF file anymore
## file.remove("afigure.pdf")
## ## Opening the stored PDF with Evince document viewer
## repo$sys("afigure.pdf", "evince")
## }
}

