% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_tags}
\alias{repo_tags}
\title{List all tags}
\usage{
repo_tags(repo)
}
\arguments{
\item{repo}{An object of class repo.}
}
\value{
Character vector of unique tags defined in the repo.
}
\description{
List all tags
}

