% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{repository}
\alias{repository}
\title{Creates a new repository backed by the given storage.}
\usage{
repository(store)
}
\arguments{
\item{store}{Read and write objects into this storage.}
}
\description{
Creates a new repository backed by the given storage.
}
