% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R, R/repository-impl.R
\name{repository_update}
\alias{repository_update}
\alias{strip_object}
\title{Removes references to environments.}
\usage{
repository_update(repo, env, plot, expr)

strip_object(obj)
}
\arguments{
\item{obj}{Object to be processed.}
}
\value{
\code{obj} with environment references replaced by \code{emptyenv()}
}
\description{
Some objects (e.g. formula, lm) store references to environments
in which they were created. This function replaces each such reference
with a reference to \code{emptyenv()}.
}
\details{
As much as possible, this function tries not to make any copies of
the original data. This is because the address of the object might
be used to determine whether object's identifier needs to be computed
which might be a costly operation.
}
