% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact.R
\name{replot}
\alias{replot}
\title{Re-plot the archived plot.}
\usage{
replot(x)
}
\arguments{
\item{x}{plot artifact, as returned by \code{\link[=read_artifacts]{read_artifacts()}}.}
}
\description{
Restore the state of the parent R session and re-run the expression
that created the given plot in that restored R session.
}
