% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_repository}
\alias{sample_repository}
\title{Work with a sample artifact repository.}
\usage{
sample_repository()
}
\description{
Returns a handle to a read/write copy of an artifact repository
installed with the package.
}
