% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{iris_models}
\alias{iris_models}
\alias{london_meters}
\title{Work with a sample artifact repository.}
\usage{
iris_models()

london_meters()
}
\value{
a \link{repository} object.
}
\description{
Obtain a handle to a read/write copy of an artifact
repository installed with the package.
}
\details{
\code{iris_models} shows a simple modeling excercise with the
\link[datasets:iris]{datasets::iris} data set.

\code{london_meters} documents a simple analysis of a subset of
the \strong{London meters} data; see the
\href{https://www.kaggle.com/jeanmidev/smart-meters-in-london}{Kaggle website}
for more details on this data set.
}
