% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transient.R
\docType{data}
\name{id}
\alias{id}
\alias{artifact}
\title{Selection helpers.}
\format{An object of class \code{NULL} of length 0.}
\usage{
id

artifact
}
\description{
Selection helpers.
}
\keyword{datasets}
