% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transient.R
\docType{data}
\name{time}
\alias{time}
\alias{object}
\alias{id}
\alias{artifact}
\alias{parent_commit}
\alias{parents}
\alias{session}
\title{Tag names.}
\format{An object of class \code{NULL} of length 0.}
\usage{
time

object

id

artifact

parent_commit

parents

session
}
\description{
Tag names need to be declared as objects to silence \code{R CMD check}.
}
\keyword{datasets}
