% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{top_n}
\alias{top_n}
\title{Retrieve top \code{n} elements.}
\usage{
top_n(.data, n, wt)
}
\arguments{
\item{.data}{A \code{query}, \code{repository} or a class supported by \code{dplyr}.}

\item{n}{The number of elements to retrieve.}

\item{wt}{(Optional). The variable to use for ordering. Supported only
in \code{dplyr}.}
}
\description{
\code{dplyr}'s \link[dplyr:top_n]{dplyr::top_n} is a function, here we define an S3 method
whose default implementation calls the function from \code{dplyr}.
}
