% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{traverse}
\alias{traverse}
\title{Traverse a graph.}
\usage{
traverse(graph, start, neighbours)
}
\arguments{
\item{graph}{graph structure, as returned by \link{connect_artifacts}.}

\item{start}{a \code{vector} of \code{character} node identifiers.}

\item{neighbours}{a \code{function} which accepts a node identifier
and the \code{graph} structure and returns a \code{vector} of node
identifiers.}
}
\value{
a \code{vector} of node identifiers visited on the traverse.
}
\description{
Traverse the whole graph or its subset. \code{graph} is a graph structure
as returned by \link{connect_artifacts}; it is a named \code{list} of \code{list}s,
each of which contains two keys: \code{children} and \code{parents}, which are
\code{character} vectors with node identifiers matching \code{names(graph)}.
The traverse starts in each identifier passed in \code{start} (there can
by multiple starting nodes). It uses \code{neighbours} to pick nodes to go
to from any given node; this function accepts two arguments: \code{id}, a
node identifier and \code{graph} which is the original graph structure.
}
\examples{
\dontrun{
   g <- ancestry_graph(...)
   # descend in a tree which start in the first node in the graph
   traverse(g, names(g)[[1]], function(id, graph) graph[[id]]$children)
}
}
\seealso{
ancestry_graph, adjust_ancestry
}
