% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact.R
\name{new_artifact}
\alias{new_artifact}
\alias{as_artifact}
\title{Create an artifact DTO.}
\usage{
new_artifact(id, store)

as_artifact(tags)
}
\arguments{
\item{id}{artifact identifier in \code{store}.}

\item{store}{Object store; see \link[storage:object_store]{storage::object_store}.}

\item{tags}{list of tag values that describe an artifact; typically
read with \code{\link[storage:os_read_tags]{storage::os_read_tags()}}.}
}
\value{
An \code{artifact} object.
}
\description{
Creates a DTO (data transfer object) that fully describes an artifact
from the repository. It is the central object for external (as opposed
to internal to this package) processing, printing, etc.
}
\details{
Each artifact (a \code{list}) has the following attributes (names):
\itemize{
\item \code{id} identifier in the object store; see \link[storage:object_store]{storage::object_store}
\itemize{
\item \code{class} one or more \code{character} values
\item \code{parents} zero or more identifiers of direct parent artifacts
\item \code{description} type-specific text describing the artifact
\item \code{expression} pre-formatted expression that produced the artifact
}
}
}
