% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact.R
\name{as_artifact}
\alias{as_artifact}
\title{Create an artifact DTO.}
\usage{
as_artifact(tags)
}
\arguments{
\item{tags}{list of tag values that describe an artifact; typically
read with \code{\link[storage:os_read_tags]{storage::os_read_tags()}}.}
}
\value{
An \code{artifact} object.
}
\description{
Creates a DTO (data transfer object) that fully describes an artifact
from the repository. It is the central object for external (as opposed
to internal to this package) processing, printing, etc.
}
