% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact.R
\name{is_artifact}
\alias{is_artifact}
\alias{is_valid_artifact}
\alias{artifact_is}
\title{Manipulating and processing artifacts.}
\usage{
is_artifact(x)

is_valid_artifact(x)

artifact_is(x, what)
}
\arguments{
\item{x}{object to be tested; \code{artifact} to be processed.}

\item{what}{property of an \code{artifact} to be tested.}
}
\description{
Manipulating and processing artifacts.

\code{artifact_is} answers various questions about an
artifact.
}
\examples{
\dontrun{
artifact_is(a, 'plot')
}
}
