% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{commit}
\alias{commit}
\alias{is_commit}
\alias{is_valid_commits}
\alias{print.commit}
\alias{commit_data}
\title{Commit API.}
\usage{
is_commit(x)

is_valid_commits(x)

\method{print}{commit}(x, ...)

commit_data(x)
}
\arguments{
\item{x}{commit object.}
}
\description{
Commit API.
}
