% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{connect_artifacts}
\alias{connect_artifacts}
\title{Turn a container of artifacts into a graph of artifacts.}
\usage{
connect_artifacts(artifacts)
}
\arguments{
\item{artifacts}{\code{container} of artifacts; see \link{read_artifacts}}
}
\value{
The input container, \code{artifacts}, whose elements are altered
as described above.
}
\description{
\code{connect_artifacts} can be called on the output of \link{read_artifacts}
to ensure that all artifacts become connected and can be turned into
a tree by calling \link{stratify}. It performs three operations:
}
\details{
\itemize{
\item given the full graph of artifacts, it connects artifacts that
are on a path in the graph (tree) but the intermediate nodes
(artifacts) are not present in the container; this produces
a \strong{connected graph}
\itemize{
\item adds a \code{children} key to each element of the container
\item removes references to parents that cannot be reassigned to
artifacts in the containers because no artifact until the very
root of the full tree is present in the container
}
}
}
