% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{history_ancestors}
\alias{history_ancestors}
\alias{history_leaves}
\alias{history_match}
\alias{history_data}
\title{Show ancestors.}
\usage{
history_ancestors(x, id)

history_leaves(x)

history_match(x, m)

history_data(x, id)
}
\description{
Show ancestors.
}
