% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transient.R
\name{ancestor_of}
\alias{ancestor_of}
\alias{no_parent}
\alias{branch_tip}
\alias{data_matches}
\title{Selectors for history queries.}
\usage{
ancestor_of(x)

no_parent()

branch_tip()

data_matches(..., data)
}
\arguments{
\item{x}{identifier of a commit or an artifact.}

\item{...}{named objects to look for.}

\item{data}{(Optional). \code{...} as a single \code{list}.}
}
\description{
Selectors for history queries.

\code{data_matches} searches for a commit whose artifacts
match objects passed via \code{...} and \code{data}.
}
