% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R, R/transient.R
\name{print.history}
\alias{print.history}
\alias{filter.history}
\alias{ancestor_of}
\alias{no_parent}
\title{History of R session(s).}
\usage{
\method{print}{history}(x, ...)

\method{filter}{history}(.data, ...)

ancestor_of(x)

no_parent()
}
\arguments{
\item{x}{History object to be printed.}

\item{...}{Extra arguments.}

\item{.data}{\code{history} object.`}
}
\description{
History of R session(s).
}
