% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R, R/transient.R
\name{commit_checkout}
\alias{commit_checkout}
\alias{ancestor_of}
\alias{no_parent}
\alias{branch_tip}
\alias{data_matches}
\title{Selectors for history queries.}
\usage{
commit_checkout(commit, env)

ancestor_of(x)

no_parent()

branch_tip()

data_matches(..., data)
}
\arguments{
\item{commit}{as returned by \code{\link[=repository_history]{repository_history()}}.}

\item{env}{Environment to check-out artifacts to.}

\item{...}{named objects to look for.}

\item{data}{(Optional). \code{...} as a single \code{list}.}
}
\description{
Selectors for history queries.

\code{data_matches} searches for a commit whose artifacts
match objects passed via \code{...} and \code{data}.
}
