% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R, R/transient.R
\name{print.history}
\alias{print.history}
\alias{filter.history}
\alias{commit_checkout}
\alias{ancestor_of}
\alias{no_parent}
\alias{branch_tip}
\alias{data_matches}
\title{History of R session(s).}
\usage{
\method{print}{history}(x, ...)

\method{filter}{history}(.data, ...)

commit_checkout(commit, env)

ancestor_of(x)

no_parent()

branch_tip()

data_matches(..., data)
}
\arguments{
\item{x}{History object to be printed.}

\item{...}{named objects to look for.}

\item{.data}{\code{history} object.`}

\item{commit}{as returned by \code{\link[=repository_history]{repository_history()}}.}

\item{env}{Environment to check-out artifacts to.}

\item{data}{(Optional). \code{...} as a single \code{list}.}
}
\description{
History of R session(s).

Selectors for history queries.

\code{data_matches} searches for a commit whose artifacts
match objects passed via \code{...} and \code{data}.
}
