% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{replot_as}
\alias{replot_as}
\alias{png_equal}
\title{Returns a base64-encoded, SVG or PNG plot.}
\usage{
replot_as(pl, type, ...)

png_equal(a, b)
}
\arguments{
\item{pl}{Plot recorded by \code{\link[=recordPlot]{recordPlot()}}.}

\item{type}{Output format: \code{"png"} or \code{"svg"}.}

\item{...}{Passed to \code{\link[=svg]{svg()}} or \code{\link[=png]{png()}}.}
}
\value{
\code{character} string, base64-encoded SVG or PNG plot.
}
\description{
Returns a base64-encoded, SVG or PNG plot.
}
