% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{print.artifact.set}
\alias{print.artifact.set}
\title{Pretty-print sets of artifacts.}
\usage{
\method{print}{artifact.set}(x, ..., style = "by_time")
}
\arguments{
\item{x}{artifact set, e.g. returned by \link{repository_explain}.}

\item{...}{further arguments passed to or from other methods.}

\item{style}{either \code{"by_time"} or \code{"tree"}.}
}
\description{
Pretty-print sets of artifacts.
}
