% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{print.repo}
\alias{print.repo}
\title{Show a summary of the repository contents.}
\usage{
\method{print}{repo}(repo, tags = NULL, all = F)
}
\arguments{
\item{repo}{An object of class repo.}

\item{tags}{A list of character tags. Only items matching all the
tags will be shown.}

\item{all}{Show also items tagged with "hide".}
}
\value{
Used for side effects.
}
\description{
Show a summary of the repository contents.
}

