% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{print.query}
\alias{print.query}
\alias{tag_names}
\alias{tag_values}
\alias{unselect}
\alias{execute}
\title{Querying an artifact repository.}
\usage{
\method{print}{query}(x, ..., simplify = FALSE)

tag_names(x)

tag_values(x)

unselect(.data)

execute(x, .warn = TRUE)
}
\arguments{
\item{x}{\code{query} object.}

\item{...}{further arguments passed to or from other methods.}

\item{simplify}{simplify output formatting.}

\item{.data}{\code{query} object.}

\item{.warn}{print warnings.}
}
\description{
Querying an artifact repository.

\code{tag_names} returns all tag names occuring among objects
selected in the query \code{x}.

\code{tag_values} returns a \code{list} of vectors of values for
all tags occuring among objects selected in the query \code{x}.

\code{unselect} clears the list of selected tag names.

\code{execute} runs the query and retrieves its results.
}
