% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R, R/selection.R
\name{is_query}
\alias{is_query}
\alias{print.query}
\alias{query}
\alias{filter.query}
\alias{select.query}
\alias{unselect}
\alias{arrange.query}
\alias{top_n.query}
\alias{summarise.query}
\alias{execute}
\alias{update}
\alias{tag_names}
\alias{tag_values}
\title{Query the repository of artifacts.}
\usage{
is_query(x)

\method{print}{query}(x, ...)

\method{filter}{query}(.data, ...)

\method{select}{query}(.data, ...)

unselect(.data)

\method{arrange}{query}(.data, ...)

\method{top_n}{query}(.data, n, wt)

\method{summarise}{query}(.data, ...)

execute(.data)

update(.data, ...)

tag_names(x)

tag_values(x)
}
\arguments{
\item{x}{Object to be tested or printed.}

\item{...}{further arguments passed to or from other methods.}

\item{.data}{\code{query} object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"query"}.
}
\description{
Query the repository of artifacts.

\code{unselect} clears the list of selected tag names.

\code{execute} runs the query and retrieves its results.

\code{tag_names} returns all tag names occuring among objects
selected in the query \code{x}.

\code{tag_values} returns a \code{list} of vectors of values for
all tags occuring among objects selected in the query \code{x}.
}
