% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{as_query}
\alias{as_commits}
\alias{as_artifacts}
\alias{is_query}
\alias{print.query}
\alias{query}
\alias{filter.query}
\alias{select_query}
\alias{unselect}
\alias{arrange.query}
\alias{top_n.query}
\alias{summarise.query}
\alias{read_artifacts}
\alias{read_commits}
\alias{execute}
\alias{update}
\title{Query the repository of artifacts.}
\usage{
as_query(x)

as_commits(x)

as_artifacts(x)

is_query(x)

\method{print}{query}(x, ...)

\method{filter}{query}(.data, ...)

select_query(.data, ...)

unselect(.data)

\method{arrange}{query}(.data, ...)

\method{top_n}{query}(.data, n, wt)

\method{summarise}{query}(.data, ...)

read_artifacts(.data)

read_commits(.data)

execute(.data)

update(.data, ...)
}
\arguments{
\item{x}{Object to be tested (\code{is_query()}, \code{is_artifacts()}, \code{is_commits()}),
printed, cast as \code{query} (\code{as_query()}, \code{as_artifacts()},
\code{as_commits()}) or querried (verbs).}

\item{...}{further arguments passed to or from other methods.}

\item{.data}{\code{query} object.}

\item{n}{The number of elements to retrieve.}

\item{wt}{(Optional). The variable to use for ordering. Supported only
in \code{dplyr}.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"query"}.
}
\description{
Query the repository of artifacts.

\code{as_query} creates a general \code{query}.

\code{as_commits} creates a \code{query} to search for commits.

\code{as_commits} creates a \code{query} to search for commits.

\code{unselect} clears the list of selected tag names.

\code{execute} runs the query and retrieves its results.
}
