% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R, R/query-specialized.R
\name{query}
\alias{query}
\alias{as_query}
\alias{is_query}
\alias{format.query}
\alias{print.query}
\alias{filter.query}
\alias{arrange.query}
\alias{top_n.query}
\alias{summarise.query}
\alias{update}
\alias{as_commits}
\alias{as_artifacts}
\alias{as_tags}
\alias{read_artifacts}
\alias{read_tags}
\alias{read_commits}
\title{Query the repository of artifacts.}
\usage{
as_query(x)

is_query(x)

\method{format}{query}(x, indent = "  ", ...)

\method{print}{query}(x, ...)

\method{filter}{query}(.data, ...)

\method{arrange}{query}(.data, ...)

\method{top_n}{query}(.data, n, wt)

\method{summarise}{query}(.data, ...)

update(.data, ...)

as_commits(x)

as_artifacts(x)

as_tags(x)

read_artifacts(.data)

read_tags(.data, ...)

read_commits(.data)
}
\arguments{
\item{x}{Object to be tested (\code{is_query()}, \code{is_artifacts()}, \code{is_commits()}),
printed, cast as \code{query} (\code{as_query()}, \code{as_artifacts()},
\code{as_commits()}) or querried (verbs).}

\item{indent}{string prepended to each line.}

\item{...}{Arguments passed on to \code{base::format}
\describe{
  \item{x}{any \R object (conceptually); typically numeric.}
}}

\item{.data}{\code{query} object.}

\item{n}{The number of elements to retrieve.}

\item{wt}{(Optional). The variable to use for ordering. Supported only
in \code{dplyr}.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"query"}.
}
\description{
\code{as_query} creates a general \code{query}.

\code{as_commits} creates a \code{query} to search for commits.

\code{as_artifacts} creates a \code{query} to search for artifacts.

\code{as_tags} creates a \code{query} to search for tag values.
}
