% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rawplot}
\alias{rawplot}
\alias{print.rawplot}
\alias{plot.rawplot}
\alias{replot_as}
\alias{as_png}
\alias{as_svg}
\alias{png_equal}
\alias{for_store}
\title{Container for plots.}
\usage{
rawplot(rp)

\method{print}{rawplot}(x, ...)

\method{plot}{rawplot}(x, ...)

replot_as(pl, type, ...)

as_png(x, width = 1280, height = 720)

as_svg(x)

png_equal(a, b)

for_store(x, store)
}
\arguments{
\item{...}{Passed to \code{\link[=svg]{svg()}} or \code{\link[=png]{png()}}.}

\item{pl}{Plot recorded by \code{\link[=recordPlot]{recordPlot()}}.}

\item{type}{Output format: \code{"png"} or \code{"svg"}.}
}
\value{
\code{character} string, base64-encoded SVG or PNG plot.
}
\description{
Container for plots.

Returns a base64-encoded, SVG or PNG plot.
}
