% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rawplot}
\alias{rawplot}
\alias{print.rawplot}
\alias{plot.rawplot}
\alias{png_equal}
\alias{for_store}
\title{Container for plots.}
\usage{
rawplot(rp)

\method{print}{rawplot}(x, ...)

\method{plot}{rawplot}(x, ...)

png_equal(a, b)

for_store(x)
}
\arguments{
\item{rp}{value returned by \link[grDevices:recordPlot]{grDevices::recordPlot}.}

\item{x}{\code{rawplot} object}

\item{...}{further arguments passed to or from other methods.}

\item{a}{PNG plot.}

\item{b}{PNG plot.}
}
\description{
Container for plots.

\code{for_store} prepares a \code{rawplot} for writing to an
object store (e.g. see \link[storage:filesystem]{storage::filesystem}).
}
