\name{repo-package}
\alias{repo-package}
\docType{package}
\title{
Getting started with repo
}
\description{
  In order to create a demo repo, use:
  
  \code{
    rp <- repo_test()    
  }
  
  This will create a test repository in your working directory. Show the
  repo content using the S3 print method:
  
  \code{
    rp
  }
  
  Repo supports 3 different interface styles. My favourite is:
  
  \code{
    rp$<function_name>(...)
  }
  
  However, for consistency with R development tools, S3-like style is
  also supported:
  
  \code{
    repo_<function_name>(rp, ...)
  }
  
  Finally, you can use "handlers". Handlers are functions generated by
  repo having the same name of the items stored in the repo. The
  function handlers retrieves a the list of such functions:

  
  \preformatted{
    h <- rp$handlers()
    item1 <- h$item1()  # loads item "item1"
    h$item1("info")     # calls the "info" function on "item1"
  }
}

