\name{repo-package}
\alias{repo-package}
\docType{package}
\title{
Getting started with repo
}
\description{
  Repo is an R objects resource manager. Use this Section to get started
  with the package.
}

\details{

  In order to create a demo repository, use:
  
  \code{
    rp <- repo_test()    
  }
  
  This will create a test repository in your working directory. Show the
  repo content using the S3 print method:
  
  \code{
    rp
  }

  Repo supports 3 different interface styles. My favourite is:
  
  \code{
    rp$<function_name>(...)
  }
  
  However, for consistency with R development tools, S3-like style is
  also supported:
  
  \code{
    repo_<function_name>(rp, ...)
  }
  
  Finally, you can use "handlers". Handlers are functions generated by
  repo having the same name of the items stored in the repo. The
  function handlers retrieves a the list of such functions:
  \preformatted{
    h <- rp$handlers() item1 <- h$item1() # loads item "item1"
    h$item1("info") # calls the "info" function on "item1" } To
  start a new empty repository, use:

  \code{
    rp <- repo_open()
  }

  To add objects to the repo, use:  \preformatted{
    x <- 1:10  # An arbitrary R object
    rp$put(x, "AnItem", "This is an example item", "temp")}
  To retreive the object later, use:
  
  \code{
    x <- rp$get("AnItem")
  }

  Look through this document for all the features currently supported by repo. Enjoy.
}


