% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\docType{package}
\name{repo-package}
\alias{repo}
\alias{repo-package}
\title{repo: The R objects resource manager}
\description{
repo: The R objects resource manager
}
\details{
The repo package is meant to help with the management of R data
files. It builds one (or more) centralized repository where R
objects are stored together with corresponding annotations, tags,
dependency notes, provenance traces. It also provides navigation
tools to easily locate and load previously stored resources.

For a complete list of functions, use \code{library(help = "repo")}.

Create a new repository with \code{repo_open()}.
}
\author{
Francesco Napolitano \email{franapoli@gmail.com}
}

