% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoS3.R
\docType{package}
\name{repo-package}
\alias{repo}
\alias{repo-package}
\title{Repo: The Data-centered Data Flow Manager}
\description{
Repo: The Data-centered Data Flow Manager
}
\details{
The Repo package is meant to help with the management of R data
files. It builds one (or more) centralized repository where R
objects are stored together with corresponding annotations, tags,
dependency notes, provenance traces. It also provides navigation
tools to easily locate and load previously stored resources.

Create a new repository with \code{rp <- repo_open()}.

Given the object rp of class \code{repo}, the \code{repo} function
\code{foo} can be called in two ways: 1) \code{repo_foo(rp)} or 2)
\code{rp$foo()}.

For a complete list of functions, use \code{library(help = "repo")}.
}
\author{
Francesco Napolitano \email{franapoli@gmail.com}
}

