% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_append}
\alias{repo_append}
\title{Append text to an existing item content.}
\usage{
repo_append(repo, id, txtorfunc)
}
\arguments{
\item{repo}{An object of class repo.}

\item{id}{The name of an item whose object is of class character.}

\item{txtorfunc}{Text to be appended to the item's object. It can
also be a on object of class function: in this case, its source is
appended.}
}
\value{
Used for side effects.

## wiping temporary repo
unlink(repo_path, TRUE)
}
\description{
This feature is experimental.
}

