% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_check}
\alias{repo_check}
\title{Check repo's integrity.}
\usage{
repo_check(repo)
}
\arguments{
\item{repo}{An object of class repo.}
}
\value{
Used for side effects.
}
\description{
Check repo's integrity.
}

