% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_copy}
\alias{repo_copy}
\title{Copy resource to another repo}
\usage{
repo_copy(repo, destrepo, name, tags = NULL)

repo_copy(repo, destrepo, name, tags = NULL)
}
\arguments{
\item{repo}{An object of class repo.}

\item{destrepo}{}

\item{name}{}

\item{tags}{}

\item{repo}{An object of class repo.}
}
\value{
Used for side effects.

Used for side effects.
}
\description{
Copy resource to another repo

Copy items to another repo
}
\examples{
###code
}

