% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_copy}
\alias{repo_copy}
\title{Copy items to another repo}
\usage{
repo_copy(repo, destrepo, name, tags = NULL)
}
\arguments{
\item{repo}{An object of class repo (will copy from it)}

\item{destrepo}{An object of class repo (will copy to it)}

\item{name}{The name of the object to copy}

\item{tags}{If not NULL, copy items matching tags.}
}
\value{
Used for side effects.
}
\description{
Copy items to another repo
}
\examples{
repo <- repo_open("_temp_repo", TRUE)
repo$put(0, "item1", "A sample item", "repo_check", replace=TRUE)
repo2 <- repo_open("_temp_repo2", TRUE)
repo$copy(repo2, "item1")
}

