% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_dependencies}
\alias{repo_dependencies}
\title{Build and/or plots the repo's dependency graph}
\usage{
repo_dependencies(repo, depends = T, attached = T, generated = T,
  plot = T)
}
\arguments{
\item{repo}{An object of class repo.}

\item{depends}{If TRUE, show "depends on" edges.}

\item{attached}{If TRUE, show "attached to" edges.}

\item{generated}{If TRUE, show "generated by" edges.}

\item{plot}{If TRUE (default), plot the dependency graph.}
}
\value{
Adjacency matrix representing the graph, with edges labeled
1, 2, 3 corresponding to "depends", "attached" and "generated" respectively.
}
\description{
Creates a weighted adjacency matrix, in which `(i,j) = x` means that
item \code{i} is in relation \code{x} with item \code{j}. The resulting graph is
plotted.
}
\details{
The relation between any two items \code{i} and \code{j} can have
    values 1, 2 or 3, respectively meaning:
\itemize{
  \item{depends on}{to build item \code{i}, item \code{j} was necessary.}
  \item{attached to}{item \code{i} is an attachment item and is attached to
        item \code{j}.}
  \item{generated by}{item \code{i} has been generated by item \code{j}. Item
        \code{j} is usually an attachment containing the source code that
        generated item \code{i}.}
}
}
\examples{
## Repository creation (or opening, if exists)
repo_path <- file.path(tempdir(), "example_repo")
repo <- repo_open(repo_path, TRUE)

## Producing some irrelevant data
data1 <- 1:10
data2 <- data1 * 2
data3 <- data1 / 2

## Putting the data in the database, specifying dependencies
repo$put(data1, "item1", "First item", "repo_dependencies", replace=TRUE)
repo$put(data2, "item2", "Item dependent on item1",
    "repo_dependencies", depends="item1", replace=TRUE)
repo$put(data3, "item3", "Item dependent on item1 and item2",
    "repo_dependencies", depends=c("item1", "item2"), replace=TRUE)

## Obtaining the dependency matrix
depmat <- repo$dependencies(plot=FALSE)
## Simplifying labels
rownames(depmat) <- colnames(depmat) <- basename(rownames(depmat))
print(depmat)
## The matrix can also be plotted as a graph (requires igraph package)
## repo$dependencies(generated=F) # turning off "generated" edges

## wiping temporary repo
unlink(repo_path, TRUE)
}

