% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_dependencies}
\alias{repo_dependencies}
\title{Build and/or plots the repo's dependency graph}
\usage{
repo_dependencies(repo, depends = T, attached = T, generated = T,
  plot = T)
}
\arguments{
\item{repo}{An object of class repo.}

\item{depends}{If TRUE, show "depends" edges}

\item{attached}{If TRUE, show "attached" edges}

\item{generated}{If TRUE, show "generated" edges}

\item{plot}{If TRUE (default), plot the dependency graph}
}
\value{
Adjacency matrix representing the graph, with edges labeled
1, 2, 3 corresponding to "depends", "attached" and "generated" edge
types respectively.
}
\description{
Build and/or plots the repo's dependency graph
}

