% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_dependencies}
\alias{repo_dependencies}
\title{Build dependency graph}
\usage{
repo_dependencies(repo, depends = T, attached = T, generated = T,
  plot = T)
}
\arguments{
\item{repo}{An object of class repo.}

\item{depends}{}

\item{attached}{}

\item{generated}{}

\item{plot}{}
}
\value{
Used for side effects.
}
\description{
Build dependency graph
}

