% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_entries}
\alias{repo_entries}
\title{Low-level list of item entries.}
\usage{
repo_entries(repo)
}
\arguments{
\item{repo}{An object of class repo.}
}
\value{
A detailed list of item entries.
}
\description{
Low-level list of item entries.
}
\examples{
repo_path <- file.path(tempdir(), "example_repo")
repo <- repo_open(repo_path, TRUE)
repo$put(1, "item1", "Sample item 1", "entries", replace=TRUE)
repo$put(2, "item2", "Sample item 2", "entries", replace=TRUE)
repo$put(3, "item3", "Sample item 3", "entries", replace=TRUE)
print(repo$entries())

## wiping temporary repo
unlink(repo_path, TRUE)
}

