% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_entries}
\alias{repo_entries}
\title{Low-level list of item entries.}
\usage{
repo_entries(rp)
}
\arguments{
\item{rp}{An object of class repo.}
}
\value{
A detailed list of item entries.
}
\description{
Low-level list of item entries.
}
\examples{
rp_path <- file.path(tempdir(), "example_repo")
rp <- repo_open(rp_path, TRUE)
rp$put(1, "item1", "Sample item 1", "entries")
rp$put(2, "item2", "Sample item 2", "entries")
rp$put(3, "item3", "Sample item 3", "entries")
print(rp$entries())

## wiping temporary repo
unlink(rp_path, TRUE)
}

