% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_open}
\alias{repo_open}
\title{Open an existing repository or create a new one.}
\usage{
repo_open(root = "~/.R_repo", forceYes = F)
}
\arguments{
\item{root}{Path to store data in.}
}
\value{
An object of class Repo.
}
\description{
Open an existing repository or create a new one.
}
\examples{
repo <- open.repo()
}

