% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_open}
\alias{repo_open}
\title{Open an existing repository or create a new one.}
\usage{
repo_open(root = "~/.R_repo", force = F)
}
\arguments{
\item{root}{Path to store data in. Defaults to "~/.R_repo".}

\item{force}{Don't ask for confirmation.}
}
\value{
An object of class Repo.
}
\description{
Open an existing repository or create a new one.
}
\examples{
## Creates a new repo in "./_temp_repo" without asking for
## confirmation.
repo <- repo_open("_temp_repo", TRUE)
repo$info()
}

