% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_root}
\alias{repo_root}
\title{Show path to repo root}
\usage{
repo_root(repo)
}
\arguments{
\item{repo}{An object of class repo.}
}
\value{
character cotaining the path to the root of the repo.
}
\description{
Show path to repo root
}
\examples{
repo_path <- file.path(tempdir(), "example_repo")
repo <- repo_open(repo_path, TRUE)
print(repo$root())

## wiping temporary repo
unlink(repo_path, TRUE)
}

