% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_public.R
\name{repo_root}
\alias{repo_root}
\title{Show path to repo root}
\usage{
repo_root()
}
\value{
character containing the path to the root of the repo.
}
\description{
Show path to repo root
}
\examples{
rp_path <- file.path(tempdir(), "example_repo")
rp <- repo_open(rp_path, TRUE)
print(rp$root())

## wiping temporary repo
unlink(rp_path, TRUE)
}
