% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_stash}
\alias{repo_stash}
\title{Quickly store temporary data}
\usage{
repo_stash(repo, name)
}
\arguments{
\item{repo}{An object of class repo.}
}
\value{
Used for side effects.
}
\description{
Quickly store temporary data
}

