% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_stashclear}
\alias{repo_stashclear}
\title{Remove all stashed data}
\usage{
repo_stashclear(repo, force = F)
}
\arguments{
\item{repo}{An object of class repo.}

\item{force}{If TRUE, no confirmation is asked.}
}
\value{
Used for side effects.
}
\description{
Remove all stashed data
}

