% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_stashclear}
\alias{repo_stashclear}
\title{Remove all stashed data}
\usage{
repo_stashclear(repo, force = F)
}
\arguments{
\item{repo}{An object of class repo.}

\item{force}{If TRUE, no confirmation is asked.}
}
\value{
Used for side effects.
}
\description{
Remove all stashed data
}
\examples{
repo_path <- file.path(tempdir(), "example_repo")
repo <- repo_open(repo_path, TRUE)
tempdata <- runif(10)
repo$stash("tempdata")
repo$print(all=TRUE)
repo$stashclear(TRUE)

## wiping temporary repo
unlink(repo_path, TRUE)
}

