% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_stashclear}
\alias{repo_stashclear}
\title{Remove all stashed data}
\usage{
repo_stashclear(rp, force = F)
}
\arguments{
\item{rp}{An object of class repo.}

\item{force}{If TRUE, no confirmation is asked.}
}
\value{
Used for side effects.
}
\description{
Remove all stashed data
}
\examples{
rp_path <- file.path(tempdir(), "example_repo")
rp <- repo_open(rp_path, TRUE)
tempdata <- runif(10)
rp$stash("tempdata")
rp$print(all=TRUE)
rp$stashclear(TRUE)

## wiping temporary repo
unlink(rp_path, TRUE)
}
\seealso{
repo_rm, repo_stash
}

