% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoS3.R
\name{repo_sys}
\alias{repo_sys}
\title{Run system call on item}
\usage{
repo_sys(repo, name, command)
}
\arguments{
\item{repo}{An object of class repo.}

\item{name}{Name of a repo item. The path to the file that contains
the item will be passed to the system program.}

\item{command}{Shell command}
}
\value{
Used for side effects.
}
\description{
Run system call on item
}

