% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-package.R, R/repository.R
\name{repository}
\alias{repository}
\alias{repository}
\alias{is_repository}
\alias{print.repository}
\alias{toString.repository}
\alias{repository_update}
\alias{repository_history}
\alias{repository_explain}
\alias{print.artifact.meta}
\alias{expl_get}
\alias{repository_rewind}
\alias{as_deltas}
\alias{as_origin}
\title{Object Repository for R}
\usage{
repository(store)

is_repository(x)

\method{print}{repository}(x, ...)

\method{toString}{repository}(x)

repository_update(repo, env, plot, expr)

repository_history(repo, mode = "all")

repository_explain(repo, id = NULL, ancestors = "unlimited")

\method{print}{artifact.meta}(x, ...)

expl_get(x, i)

repository_rewind(repo, id)

as_deltas(x)

as_origin(x)
}
\arguments{
\item{store}{Read and write objects into this storage.}

\item{x}{Object to be tested, converted or printed.}

\item{repo}{A repository object.}

\item{env}{Environment to create a commit from (e.g. \link{globalenv}).}

\item{plot}{A recorded plot (see \link[grDevices:recordPlot]{grDevices::recordPlot}).}

\item{expr}{The expression related to the most recent changed in \code{env}.}

\item{id}{Commit identifier.}

\item{ancestors}{Retrieve ancestors at most that far in the ancestry
tree from specified \code{id}s.}

\item{repo}{Repository object.}
}
\description{
Tracks full history of multiple consecutive R sessions.

Creates a new repository backed by the given storage.

\code{is_repository} verifies whether \code{x} is a repository
object.

\code{repository_update} appends a new commit to the repository.

\code{repository_history} returns a graph that describes the
execution of commands in R session, also know as the \strong{time} view.
Each node in that graph represents the state of R session (aka. a
\strong{commit}) at a given point in time. Each edge represents a single
R command issued by the user. If the optional parameter \code{id} is
specified, the history is limited to the subtree of \strong{commits}
with \code{id} as the root.

\code{repository_explain} returns a graph that describes
the \emph{origin} of an artifact (an R object or a plot) with the given
identifier \code{id}. If no \code{id} is provided, an aggregated graph containing
all artifacts is returned.

\code{print.explained} pretty-prints a description of an
artifact.

\code{expl_get} extracts a member \code{i} from the \code{"explained"}
S3 object.

\code{repository_rewind} changes the internal pointer to the
\emph{last commit} and, if \code{id} denotes a historical commit, sets it to
that value. Subsequent commits will be recorded as descendants of
commit \code{id}.

\code{as_deltas} converts a \code{history} object (a tree of
\emph{commits}) into an equivalent tree of \emph{artifacts}. Each node in the
tree of artifacts represents a single artifact introduced in a given
commit. It is still a representation of historical changes in R
session but at the level of a single artifact rather than at the
level of a snapshot of R session (a \emph{commit}).
}
