% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-package.R, R/repository.R
\name{repository}
\alias{repository}
\alias{repository}
\alias{repository_update}
\alias{repository_history}
\alias{repository_last_commit}
\alias{repository_explain}
\title{Object Repository for R}
\usage{
repository(store)

repository_update(repo, env, plot, expr)

repository_history(repo, id = NULL)

repository_last_commit(repo)

repository_explain(repo, id = NULL)
}
\arguments{
\item{store}{Read and write objects into this storage.}
}
\description{
Tracks full history of multiple consecutive R sessions.

Creates a new repository backed by the given storage.

\code{repository_history} returns a graph that describes the
execution of commands in R session, also know as the \strong{time} view.
Each node in that graph represents the state of R session (aka. a
\strong{commit}) at a given point in time. Each edge represents a single
R command issued by the user. If the optional parameter \code{id} is
specified, the history is limited to the subtree of \strong{commits}
with \code{id} as the root.

\code{repository_explain} returns a graph that describes
the \strong{origin} of an artifact (an R object or a plot) with the given
identifier \code{id}. If no \code{id} is provided, an aggregated graph containing
all artifacts is returned.
}
