% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{samples}
\alias{samples}
\alias{iris_model}
\alias{london_meters}
\alias{execute_simulation}
\title{Work with a sample artifact repository.}
\usage{
iris_model()

london_meters()

execute_simulation(path, repo)
}
\arguments{
\item{path}{path to a script to be executed.}

\item{repo}{store artifacts in this repository.}
}
\value{
a \link{repository} object.
}
\description{
Obtain a handle to a read/write copy of an artifact
repository installed with the package.
}
\details{
\code{iris_model} shows a simple modeling excercise with the
\link[datasets:iris]{datasets::iris} data set.

\code{london_meters} documents a simple analysis of a subset of
the \strong{London meters} data; see the
\href{https://www.kaggle.com/jeanmidev/smart-meters-in-london}{Kaggle website}
for more details on this data set.

\code{execute_simulation} runs a script inside a simulated R
session with a artifact tracker enabled. Artifacts are stored in
repository \code{repo}.
}
