% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{generate_simple}
\alias{generate_simple}
\alias{simulate_london_meters}
\title{Simulations and examples.}
\usage{
generate_simple(repo)

simulate_london_meters(repo)
}
\arguments{
\item{overwrite}{If current stash contains objects, setting \code{overwrite}
to \code{TRUE} will remove them prior to running simulation.}
}
\description{
\code{simulate_london_meters} loads and examines a subset
of \strong{London meters} data; see the (https://www.kaggle.com/jeanmidev/smart-meters-in-london)\link{Kaggle website}
for this data set and the introductory vignette.
}
\details{
These functions populate sessions' history cache with a complete
history of data exploration.
}
